move common handler code into handlers.py
