+++ lib/handlers/application_handler_v1.py
@@ -8,6 +8,7 @@ from patent grant documents
 from cStringIO import StringIO
 from datetime import datetime
 from unidecode import unidecode
+from handler import Patobj, PatentHandler
 import re
 import uuid
 import xml.sax
@@ -17,11 +18,7 @@ import xml_driver
 claim_num_regex = re.compile(r'^\d+\. *') # removes claim number from claim text
 
 
-class Patobj(object):
-    pass
-
-
-class Patent(object):
+class Patent(PatentHandler):
 
     def __init__(self, xml_string, is_string=False):
         xh = xml_driver.XMLHandler()
@@ -35,6 +32,9 @@ class Patent(object):
             parser.parse(StringIO(xml_string))
         else:
             parser.parse(xml_string)
+
+        self.attributes = ['app', 'application', 'assignee_list']
+
         self.xml = xh.root.us_patent_application
 
         self.country = self.xml.publication_reference.contents_of('country', upper=False)[0]
@@ -153,9 +153,3 @@ class Patent(object):
                 asg['uuid'] = str(uuid.uuid1())
                 res.append([asg, loc])
         return res
-
-    def get_patobj(self):
-        patobj = Patobj()
-        for attr in ['app', 'application', 'assignee_list']:
-            patobj.__dict__[attr] = getattr(self, attr)
-        return patobj
